Refactor App and Enhance Components for Improved User Experience
Updated the App component to include a new HomeJob route and adjusted the NavigationSticky component to accept dynamic styles. Enhanced the BusinessCard component with updated titles and copy for clarity. Refined the Intro component with improved styling and added typewriter effects for better engagement. Introduced animation effects in the Portfolio component for smoother visibility transitions and added new sections to highlight leadership principles and personal values.