Refactor Identity and Navigation Components for Improved Structure and Responsiveness
Updated the IdentitySticky and Navigation components to enhance layout and styling. Changed the Identity component from a link to a div for better structure. Adjusted the Wortmarke component to use a background image instead of an img tag, improving responsiveness across devices. Refined the NavigationSticky component to include the Wortmarke and adjusted styles for better alignment and spacing. These changes aim to enhance the overall user experience and maintainability of the components.